feat: Add logging to file, post-processing analysis, custom IDs
- Add proper logging, with output to stdout and file. Only
implemented within pipeline.py and still need to add logging from
external modules.
- Add a quick analysis after final image combination that measures
the 1-d radial profile and plots it as a check for good/bad
PSF-subtraction and possible disk signal. Next step is to add an
automatic interpretation of the profile to inform the user if the
profile indicates any of those cases.
- Add --cid option to main_reduce_stis.py to attach a custom
suffix to output filenames of a given run for file organization.