feat: Output final image without radial profile subtraction; more
- Add a new slice to the final image cube output; index 2 is now
a version with no radial profile subtraction done at all.
- Add slices to the SNR cube output that match the final image
cube now.
- Stop filling new pixels created by de-rotation with 1.0;
fill them with NaN now instead.
- Write exclusions to the header comments.
- Adjust where error maps sample the image standard deviation, to
use a wider region and to use the minor axis on the back edge of
the disk (based on the disk PA). Also sample for inner
interpolation along the Y axis now in addition to the X axis.