feat: Implement auto star masking with astrosniff and iteration
- Add auto star masking with astrosniff if a final image already
exists in the data directory. Combines with the source masks.
- Add iterative version of the pipeline with --iterate flag. This
runs the pipeline once normally and then runs it again immediately
afterwards from the axc step using the output from the first run.
This will allow auto star masking to work for brand new datasets
by producing a final image from which to make the mask in the
second run.
- Move astrosniff from scripts to improcess
- Bump version number