fix: Revise saturation masks to handle bad DQ flags; more
- Revise saturation masks based on DQ flags to only mask a pixel
if its DN value or that of at least one adjacent pixel is actually
over a saturation threshold. We conservatively set the threshold
to 0.8*27000 DN, where the latter is the lowest saturation value
for any pixel in the STIS CCD. To account for blurring of
saturation into adjacent pixels by distortion correction and
interpolation within the pipeline, we consider adjacent pixels up
to 2 pixels away in all directions.
- Add logging of dataset's temporal length in hours and full date
range of included data.
- Save PSF reference star name in FITS headers more reliably.