fix: Pad axt-level images with NaNs; maybe fix ref auto star mask
- Change image padding during image alignment so the empty
values in the large padded axt/axc image are NaN instead of zero
or something else.
- Also reorganize image shifting function to facilitate the
padding fix. Now, we do the subpixel shift first on the original
image, and then finish with the bulk whole-pixel shift afterwards.
The resulting interpolation is unchanged but the edges of the data
are slightly cleaner and padding with NaNs works now.
- Maybe fix auto star masking of reference images. It at least runs
without crashing now, but not confirmed that it works properly.