fix: (major) Correct intensity conversion to COUNTS S^-1
flt images are now correctly converted from intensity units of
COUNTS to COUNTS PER SECOND after the unification step. Previously,
the conversion was not properly accounting for the number of
CRSPLITS that were unified. This is fixed with this commit.

An attempt was also made to keep this fix backwards compatible with
sx2 reductions by basing the conversion on the inputType. This is
not tested for sx2 data.