fix (major): Intensity calibration, auto star masking, Gaia MCMC
- Correct the intensity calibration to units of COUNTS S^-1.
Previously, the unified images were being normalized by only the
exposure time of a single CRSPLIT. Now, they are normalized by the
total integration time of the image (the summed exposure times of
all CRSPLITS) to match the way the unified images conserve the
intensity in COUNTS.
- Fix auto star masking so it runs on final image and reference
PSF images.
- Actually apply the auto mask to the reference PSF during PSF
subtraction. This removes negative residuals in PSF-subtracted
science images caused by sources in the reference images, but also
opens the possibility that improper auto masking in the reference
image can mask pixels unnecessarily in the final image.
- Reduce amount of reference image's primary PSF that is masked
by the auto star masker. Do this by performing a "reverse PSF
subtraction" by subtracting a median of the science images from
the reference PSF image so we can locate and mask the stars in
the reference image without the primary PSF in the way.
- Pass more arguments to auto star masking function calls.
- Adjust Gaia astrometry MCMC prior and walker initialization to
avoid zombie walkers that are initialized outside the prior
bounds. Add 0 lower bounds for MCMC's x and y centers.
- Update Gaia PSF least-squares fitter with modern LMLSQFitter
class.
- Make Gaia overview star labels a little nicer.